requirements.txt
backtrader==1.9.76.123
dash==2.7.1
finta==1.3
Flask==2.1.0
Flask_Caching==2.0.1
Flask_Cors==3.0.10
ibapi==9.81.1.post1
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.3
plotly==5.11.0
python-dotenv==0.21.0
python_dateutil==2.8.2
pytz==2022.1
scipy==1.7.3
ta==0.10.2
python-binance==1.0.17
